Rename pgp_key -> PgpKey
